requirements.txt
langchain
openai
pypdf
faiss-cpu
streamlit
python-dotenv
tiktoken
pyttsx3
speechrecognition
pyaudio